Add unit test for UserMapper; Add multi-lang support for BadRequest response; Make some small change for API documentation; Config pageSize in PaginationResponse as number of record in this page

+++ src/main/java/shop/haui_megatech/domain/dto/pagination/PaginationResponseDTO.java
@@ -10,7 +10,7 @@ public record PaginationResponseDTO<T>(
 		@JsonInclude(JsonInclude.Include.NON_NULL)
 		String keyword,
 		Integer pageIndex,
-		Short pageSize,
+		Integer pageSize,
 		Long totalItems,
 		Integer totalPages,
 		List<T> items
